moved some code around to new files, implemented new detailed job row, implemented new detailed job selector views, fixed display bug on planning
